CATEGORIA PATH

+++ src/main/java/com/mc/blog/services/CategoriaService.java
@@ -8,13 +8,18 @@ import com.mc.blog.repositories.ArtigosRepository;
 import com.mc.blog.repositories.CategoriaRepository;
 import com.mc.blog.security.UserSS;
 import com.mc.blog.services.exceptions.AuthorizationException;
+import com.mc.blog.services.exceptions.DataIntegrityException;
 import com.mc.blog.services.exceptions.ObjectNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Example;
+import org.springframework.data.domain.ExampleMatcher;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.Arrays;
+import java.util.List;
 import java.util.Optional;
 
 @Service
@@ -22,6 +27,9 @@ public class CategoriaService {
 	
 	@Autowired
 	private CategoriaRepository repo;
+
+	@Autowired
+	private ArtigosRepository artigosRepository;
 
 	@Transactional(readOnly = true)
 	public Categoria find(Long id) {
@@ -57,7 +65,30 @@ public class CategoriaService {
 	}
 
 	public void delete(Long id) {
-		find(id);
+
+		Categoria categoria = find(id);
+		Categoria cat = new Categoria();
+		cat.setCategoriaPai(categoria);
+
+
+		Example<Categoria> example = Example.of(cat, ExampleMatcher.matching().withStringMatcher(ExampleMatcher.StringMatcher.CONTAINING)   // Match string containing pattern
+				.withIgnoreCase());
+		List<Categoria> categorias  = repo.findAll(example);
+		if(!categorias.isEmpty()){
+			throw new DataIntegrityException("Categoria possui subcategorias");
+		}
+
+		Artigos artigos = new Artigos();
+		artigos.setCategorias(Arrays.asList(categoria));
+
+		Example<Artigos> exampleArt = Example.of(artigos, ExampleMatcher.matching().withStringMatcher(ExampleMatcher.StringMatcher.CONTAINING)   // Match string containing pattern
+				.withIgnoreCase());
+		List<Artigos> artigos1  = artigosRepository.findAll(exampleArt);
+
+		if(artigos1.size()>0){
+			throw new DataIntegrityException("Categoria possui artigos relacionados");
+		}
+
 		repo.deleteById(id);
 	}
